+++ src/models/ContentFolder.ts
@@ -12,6 +12,9 @@ export interface ContentFolder {
   originalPath?: string;
   $schema?: string;
   extended?: boolean;
+
+  locale?: string;
+  localeSourcePath?: string;
   defaultLocale?: string;
   locales: I18nConfig[];
 }